Move snapshot logic
