+++ src/subgraphs/basin/src/entities/snapshots/Beanstalk.ts
@@ -1,0 +1,1 @@
+// takeBeanstalkSnapshots(beanstalk, block)
